Refactor parser runtime infrastructure files and scheduler naming
