+++ Utils.Parser/Runtime/ParserRuleResult.cs
@@ -1,0 +1,3 @@
+namespace Utils.Parser.Runtime;
+
+internal readonly record struct ParserRuleResult(ParseNode? Node, int EndPosition, bool IsFailure);
